+++ src/main/kotlin/com/yapp/itemfinder/domain/BaseEntity.kt
@@ -1,0 +1,29 @@
+package com.yapp.itemfinder.domain
+
+import com.yapp.itemfinder.common.Const.KST_ZONE_ID
+import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+import javax.persistence.Column
+import javax.persistence.EntityListeners
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+import javax.persistence.MappedSuperclass
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener::class)
+abstract class BaseEntity(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0L
+) {
+    @CreatedDate
+    @Column(updatable = false)
+    var createdAt: LocalDateTime = LocalDateTime.now(KST_ZONE_ID)
+
+    @LastModifiedDate
+    @Column(updatable = true)
+    var updatedAt: LocalDateTime = LocalDateTime.now(KST_ZONE_ID)
+}
