--- src/main/kotlin/com/yapp/itemfinder/domain/entity/item/ItemPin.kt
@@ -1,9 +1,0 @@
-package com.yapp.itemfinder.domain.entity.item
-
-import javax.persistence.Embeddable
-
-@Embeddable
-class ItemPin(
-    var positionX: Float,
-    var positionY: Float
-)
