--- src/main/kotlin/com/yapp/itemfinder/domain/entity/member/MemberRepository.kt
@@ -1,5 +1,0 @@
-package com.yapp.itemfinder.domain.entity.member
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface MemberRepository : JpaRepository<MemberEntity, Long>
