+++ src/main/kotlin/com/yapp/itemfinder/domain/item/ItemPin.kt
@@ -1,0 +1,9 @@
+package com.yapp.itemfinder.domain.item
+
+import javax.persistence.Embeddable
+
+@Embeddable
+class ItemPin(
+    var positionX: Float,
+    var positionY: Float
+)
