--- src/main/kotlin/com/yapp/itemfinder/domain/repository/TokenRepository.kt
@@ -1,6 +1,0 @@
-package com.yapp.itemfinder.domain.repository
-
-import com.yapp.itemfinder.domain.entity.token.TokenEntity
-import org.springframework.data.repository.CrudRepository
-
-interface TokenRepository : CrudRepository<TokenEntity, Long>
