+++ src/main/kotlin/com/yapp/itemfinder/domain/space/dto/CreateSpaceRequest.kt
