+++ src/main/kotlin/com/yapp/itemfinder/domain/space/dto/SpaceResponse.kt
@@ -1,0 +1,24 @@
+package com.yapp.itemfinder.domain.space.dto
+
+import com.yapp.itemfinder.domain.space.SpaceEntity
+
+data class SpaceResponse(val name: String, val id: Long) {
+    companion object {
+        fun from(space: SpaceEntity): SpaceResponse {
+            return SpaceResponse(
+                name = space.name,
+                id = space.id
+            )
+        }
+    }
+}
+
+data class SpacesResponse(val spaces: List<SpaceResponse>) {
+    companion object {
+        fun from(spaces: List<SpaceEntity>): SpacesResponse {
+            return SpacesResponse(
+                spaces = spaces.map { SpaceResponse.from(it) }
+            )
+        }
+    }
+}
